services/orquestador/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
httpx==0.27.2
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.21.0
python-json-logger==2.0.7
